Update 14 Maret

+++ LireOffice/Models/Data/SalesItem.cs
@@ -13,6 +13,6 @@
         public decimal SellPrice { get; set; }
         public decimal Discount { get; set; }
         public decimal SubTotal { get; set; }
-        public decimal Tax { get; set; }
+        public double Tax { get; set; }
     }
 }
